// games-BE-ac-v2/libs/shared/enums/redisCacheKeys.enum.ts
export enum RedisCacheKeyEnum {
  USER_DETAILS = 'userDetails',
  BET_LIST = 'betList',
  USER_BETS_ROUND = 'userBetsRound',
  ACTIVE_GAME = 'activeGame',
  TOTAL_BET_COUNT = 'totalBetCount',
  PLAYER_BET_COUNT = 'playerBetCount',
  LAST_ROUND_LIST = 'lastRoundList',
  GAME_CONFIG = 'gameConfig',
  ACTIVE_BET = 'activeBet',
}